+++ src/axom/core/Types.hpp
@@ -25,6 +25,25 @@
 
 namespace axom
 {
+using float32 = float;
+using float64 = double;
+
+/*
+  Axom integer types are deprecated.
+  Their use can trigger an warning or an error, or be quietly allowed,
+  by using CMake -DAXOM_DEPRECATED_TYPES=<WARN|ERROR|ALLOW>
+  Eventually, these types will be removed.
+*/
+#if AXOM_DEPRECATED_TYPES_N == 1 || AXOM_DEPRECATED_TYPES_N == 2
+  #if AXOM_DEPRECATED_TYPES_N == 1
+    #if defined(_MSC_VER)
+      #pragma message( \
+        "warning: Using deprecated Axom types.  Please see CMake variable AXOM_DEPRECATED_TYPES")
+    #else
+      #warning \
+        "Using deprecated Axom types.  Please see CMake variable AXOM_DEPRECATED_TYPES"
+    #endif
+  #endif
 using int8 = std::int8_t;   /*!< 8-bit signed integer type      */
 using uint8 = std::uint8_t; /*!< 8-bit unsigned integer type    */
 
@@ -34,20 +53,18 @@ using uint16 = std::uint16_t; /*!< 16-bit unsigned integer type   */
 using int32 = std::int32_t;   /*!< 32-bit signed integer type     */
 using uint32 = std::uint32_t; /*!< 32-bit unsigned integer type   */
 
-// Note: KW -- We assume that AXOM_NO_INT64_T will be defined
-// on systems/compilers that do not support 64 bit integer types
-#ifndef AXOM_NO_INT64_T
+  // Note: KW -- We assume that AXOM_NO_INT64_T will be defined
+  // on systems/compilers that do not support 64 bit integer types
+  #ifndef AXOM_NO_INT64_T
 using int64 = std::int64_t;   /*!< 64-bit signed integer type     */
 using uint64 = std::uint64_t; /*!< 64-bit unsigned integer type   */
+  #endif
 #endif
 
-using float32 = float;
-using float64 = double;
-
 #if defined(AXOM_USE_64BIT_INDEXTYPE) && !defined(AXOM_NO_INT64_T)
-using IndexType = int64;
+using IndexType = std::int64_t;
 #else
-using IndexType = int32;
+using IndexType = std::int32_t;
 #endif
 
 #ifdef AXOM_USE_MPI
@@ -83,42 +100,42 @@ struct mpi_traits<float32>
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<int8>
+struct mpi_traits<std::int8_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<uint8>
+struct mpi_traits<std::uint8_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<int16>
+struct mpi_traits<std::int16_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<uint16>
+struct mpi_traits<std::uint16_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<int32>
+struct mpi_traits<std::int32_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<uint32>
+struct mpi_traits<std::uint32_t>
 {
   static const MPI_Datatype type;
 };
@@ -126,14 +143,14 @@ struct mpi_traits<uint32>
   //------------------------------------------------------------------------------
   #ifndef AXOM_NO_INT64_T
 template <>
-struct mpi_traits<int64>
+struct mpi_traits<std::int64_t>
 {
   static const MPI_Datatype type;
 };
 
 //------------------------------------------------------------------------------
 template <>
-struct mpi_traits<uint64>
+struct mpi_traits<std::uint64_t>
 {
   static const MPI_Datatype type;
 };